Adding Movement class in Animals class.
