--- src/main/java/com/march/project/homework5/Dog.java
@@ -1,12 +1,0 @@
-package com.march.project.homework5;
-
-public class Dog extends Animals {
-    public Dog(String name) {
-        this.name = name;
-        this.runningSpeed = 10;
-        this.swimmingSpeed = 5;
-        this.powerRequirements = 2;
-        this.endurance = 20;
-        this.isEndurance = false;
-    }
-}
