--- src/main/java/com/march/project/homework5/Horse.java
@@ -1,12 +1,0 @@
-package com.march.project.homework5;
-
-public class Horse extends Animals {
-    public Horse(String name) {
-        this.name = name;
-        this.runningSpeed = 40;
-        this.swimmingSpeed = 4;
-        this.powerRequirements = 4;
-        this.endurance = 40;
-        this.isEndurance = false;
-    }
-}
